Added `from` command to quote msg from other channels
I simply copied the `id` implementation and changed the search channel. I did not test this and I don't know if any other changes could be required.
I don't know exactly what the `*` parameters are meant to be, I assumed they were for the spaces? so I added one after the channel to hopefully divide the channel from the messageID.